Add files via upload
Monitoring and Maintaining the Street light is a big task for the people of concern department 
because the network contains thousands of street lights. It is very difficult to identify the 
faulty or failed strret light. The main purpose of this project is to save the energy , due to 
continuos glowing of strret lights. GSM Technology is used to locate the faulty strret light 
such that the maintaince person can reach to the post directly to rectify the faults . The 
advantage of GSM Technology that there is no range restriction, the information can be sent 
to the authority user where the person . The person can operate through their’s Mobile.
It is energy saving process. No more energy is wasted. We can operate it any time and 
anywhere regardless of however nice a brand new application is made , there continosly are 
often thing further to it. The authority user must be responsible for the alert message and 
make the fault correctly. The IOT also helps monitoring the data in real time.
Street Light monitoring is an important application of the Internet of Things. We propose an 
Street Light monitoring system using IoT. The main objective of this model is to receive 
messages and track from any location. Real time deployment of the proposed model is to be 
carried out in the future. Maintenance of the equipment in all weather conditions, 
transmission of data effectively.